Turning off a bunch of the printlns because they disrupt the output.
Also added a note that we need to *keep* the last println in the demo.
Otherwise, the demo also misbehaves (but in a manner that I believe I
can understand, in terms of potential compiler optimizations that
might be happening).

VERY IMPORTANT: at some point soon, we must make a test that
re-enables these println's, because not only did I find that they were
disrupting the output, they were also causing the demo to misbehave. I
do not think this is solely an artfiact of the way I have hacked in
the client-request integration, because I saw problems even with the
relatively late printlns in the output.