Add new crate for UI tests
This patch adds a new crate called `kc` which is used for all Valgrind
UI tests. This makes uses of the `ui_test` crate which already exists
and is currently used for Miri.

In order to call not just rustc, but also Valgrind, this patch adds a
wrapper script that does both of those things.