Further refinement of the Makefile. Having a single recipe that always called autogen.sh, even when e.g.  just krabcake's Makefile.am changed, led to some bad behavior where it would unconditionally invoke autogen.sh because the recipe said that `configure` should be regenerated, which does *not* happen if the `configure.ac`, and probably some other dependendencies of which I'm ignorant, have not been changed.
Crucially, touching Makefile.am and then re-running autogen.sh would
leave the configure script untouched. Which is a good thing, but it
means that our Makefile needs to embed a little more of the autogen.sh
logic internally if its going to behave well and not keep invoking
autogen.sh on each run.